refactor order status
