+++ packages/kokoas-client/src/pages/order/orderRequestForm/orderDialogTitle/OrderStatus.tsx
@@ -1,28 +1,14 @@
 import { Chip, Zoom } from '@mui/material';
-import { useOrderWatch } from '../hooks/useOrderRHF';
-import { TOrderForm } from '../schema';
 import { statusBGcolorMap, statusFGcolorMap } from 'kokoas-client/src/lib/progressColors';
-import { useOrderById } from 'kokoas-client/src/hooksQuery';
-import { KProgress } from 'types/src/common/order';
+import { useOrderStatus } from '../hooks/useOrderStatus';
 
 export const OrderStatus = () => {
-  const orderId = useOrderWatch({
-    name: 'orderId',
-  }) as TOrderForm['orderId'];
-
-  const { 
-    data, 
-    isFetching,
-  } = useOrderById({ orderId });
 
 
   const {
-    status,
-  } = data || {};
-
-  const orderStatus = (orderId 
-    ? status?.value || '未発注'
-    : '新規') as KProgress | undefined;
+    isFetching,
+    orderStatus,
+  } = useOrderStatus();
 
   return (
     <Zoom in={!isFetching && !!orderStatus} timeout={2000}>
